+++ src/box/journal.c
@@ -33,6 +33,7 @@
 #include <diag.h>
 #include "error.h"
 #include "watcher.h"
+#include "xrow.h"
 
 struct journal *current_journal = NULL;
 
@@ -122,4 +123,22 @@ journal_queue_flush(void)
 	while (!rlist_empty(list))
 		fiber_wakeup(rlist_first_entry(list, struct fiber, state));
 	journal_queue_wait();
+}
+
+int
+journal_write_row(struct xrow_header *row)
+{
+	char buf[sizeof(struct journal_entry) + sizeof(struct xrow_header *)];
+	struct journal_entry *entry = (struct journal_entry *)buf;
+	entry->rows[0] = row;
+	journal_entry_create(entry, 1, xrow_approx_len(row),
+			     journal_entry_fiber_wakeup_cb, fiber());
+
+	if (journal_write(entry) != 0)
+		return -1;
+	if (entry->res < 0) {
+		diag_set_journal_res(entry->res);
+		return -1;
+	}
+	return 0;
 }
